v1.0.0 Recoded
I decided as i've completed the basic project for Kick4J that I would go ahead and recode the full project and tidy up the code.

With the odd rate limit here and there like usual (v1.0.0 had the same) it is working so much more efficiently and with a LOT less lines of code which just pleases the eyes of any programmer.

Updates:
 - Removed Chatroom feature, it's not really needed. If you would like it added back then just go ahead and mention it in the issues tab (https://github.com/MistyKnives/Kick4J/issues) for me and i'll happily add it back :)
 - After a few hours of research on the Jackson API that I use to parse all my JSON data, I have removed all the useless lines of code and made it much more flawless.
 - All the updated methods will be in the README.md file located at the front page of the github project (https://github.com/MistyKnives/Kick4J/blob/main/README.md).

I am extremely pleased with how this turned out. Huge shoutout to the makers of the Jackson API for an extremely useful plugin. Could not have done it without it.