Kick4J v1.0 - Recoded
Hello,
Welcome to the new and improved Kick4J. From a small little Website Scraper API. We are now a full blown Client Library. With Support for Commands, Events, API, and more! We've worked very hard to bring this to you so there may be bugs, issues, etc. Please report them straight away and we will work on a fix as soon as possible.

Documentation is on its way - https://docs.kick4j.com/

Enjoy!